merge: hack the index more throghrouly to avoid all racy protections
Change the timestamps such that we never hit the racy protections. We
first make the index think that the modified entries do not have racy
timestamps so we don't detect that it's changed.

We then hack the timestamp some more to make the index think we do not
have the same timestamp as the entry such that we do not look at the
file's content and detect that they have changed.